+++ earth_extractor/satellites/sentinel.py
@@ -1,6 +1,6 @@
 from earth_extractor.satellites.base import Satellite
 from earth_extractor.satellites import enums
-from earth_extractor.providers import copernicus_scihub, asf
+from earth_extractor.providers import copernicus_dataspace, asf
 from earth_extractor import core
 import logging
 
@@ -13,7 +13,7 @@ logger.setLevel(core.config.constants.LOGLEVEL_MODULE_DEFAULT)
 """
 
 sentinel_1: Satellite = Satellite(
-    query_provider=copernicus_scihub,
+    query_provider=copernicus_dataspace,
     download_provider=asf,
     name=enums.Satellite.SENTINEL1,
     description="Sentinel 1",
@@ -23,8 +23,8 @@ sentinel_1: Satellite = Satellite(
 )
 
 sentinel_2: Satellite = Satellite(
-    query_provider=copernicus_scihub,
-    download_provider=copernicus_scihub,
+    query_provider=copernicus_dataspace,
+    download_provider=copernicus_dataspace,
     name=enums.Satellite.SENTINEL2,
     description="Sentinel 2",
     processing_levels=[enums.ProcessingLevel.L1C, enums.ProcessingLevel.L2A],
@@ -33,8 +33,8 @@ sentinel_2: Satellite = Satellite(
 )
 
 sentinel_3: Satellite = Satellite(
-    query_provider=copernicus_scihub,
-    download_provider=copernicus_scihub,
+    query_provider=copernicus_dataspace,
+    download_provider=copernicus_dataspace,
     name=enums.Satellite.SENTINEL3,
     description="Sentinel 3",
     processing_levels=[enums.ProcessingLevel.L1, enums.ProcessingLevel.L2],